Moved parsePositionFromString.ts to utils and restyled all the functions there to upcoming reusing them in api
